backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# ML & AI
scikit-learn==1.3.2
xgboost==2.0.2
pandas==2.1.4
numpy==1.25.2
shap==0.43.0
lime==0.2.0.1
scipy>=1.7.0

# Data Processing
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# File Processing
aiofiles==23.2.1
openpyxl==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Environment
python-dotenv==1.0.0
psutil==5.9.6
